+++ pyteal/compiler/compiler.py
@@ -1,10 +1,13 @@
-from typing import List, DefaultDict, cast
-from collections import defaultdict
+from typing import List
 
-from .ast import Expr
-from .ir import Op, Mode, TealComponent, TealOp, TealLabel, TealBlock, TealSimpleBlock, TealConditionalBlock
-from .errors import TealInputError, TealInternalError
-from .config import NUM_SLOTS
+from ..ast import Expr
+from ..ir import Mode, TealComponent, TealOp, TealBlock
+from ..errors import TealInputError, TealInternalError
+from ..config import NUM_SLOTS
+
+from .sort import sortBlocks
+from .flatten import flattenBlocks
+from .constants import createConstantBlocks
 
 MAX_TEAL_VERSION = 3
 MIN_TEAL_VERSION = 2
@@ -15,95 +18,6 @@ class CompileOptions:
     def __init__(self, *, mode: Mode = Mode.Signature, version: int = DEFAULT_TEAL_VERSION):
         self.mode = mode
         self.version = version
-
-def sortBlocks(start: TealBlock) -> List[TealBlock]:
-    """Topologically sort the graph which starts with the input TealBlock.
-
-    Args:
-        start: The starting point of the graph to sort.
-
-    Returns:
-        An ordered list of TealBlocks that is sorted such that every block is guaranteed to appear
-        in the list before all of its outgoing blocks.
-    """
-    # based on Kahn's algorithm from https://en.wikipedia.org/wiki/Topological_sorting
-    S = [start]
-    order = []
-
-    while len(S) != 0:
-        n = S.pop(0)
-        order.append(n)
-        for i, m in enumerate(n.getOutgoing()):
-            for i, block in enumerate(m.incoming):
-                if n is block:
-                    m.incoming.pop(i)
-                    break
-            if len(m.incoming) == 0:
-                if i == 0:
-                    S.insert(0, m)
-                else:
-                    S.append(m)
-    
-    return order
-
-def flattenBlocks(blocks: List[TealBlock]) -> List[TealComponent]:
-    """Lowers a list of TealBlocks into a list of TealComponents.
-
-    Args:
-        blocks: The blocks to lower.
-    """
-    codeblocks = []
-    references: DefaultDict[int, int] = defaultdict(int)
-
-    indexToLabel = lambda index: "l{}".format(index)
-
-    for i, block in enumerate(blocks):
-        code = list(block.ops)
-        codeblocks.append(code)
-        if block.isTerminal():
-            continue
-
-        if type(block) is TealSimpleBlock:
-            simpleBlock = cast(TealSimpleBlock, block)
-            assert simpleBlock.nextBlock is not None
-
-            nextIndex = blocks.index(simpleBlock.nextBlock, i+1)
-            if nextIndex != i + 1:
-                references[nextIndex] += 1
-                code.append(TealOp(None, Op.b, indexToLabel(nextIndex)))
-        elif type(block) is TealConditionalBlock:
-            conditionalBlock = cast(TealConditionalBlock, block)
-            assert conditionalBlock.trueBlock is not None
-            assert conditionalBlock.falseBlock is not None
-
-            trueIndex = blocks.index(conditionalBlock.trueBlock, i+1)
-            falseIndex = blocks.index(conditionalBlock.falseBlock, i+1)
-
-            if falseIndex == i + 1:
-                references[trueIndex] += 1
-                code.append(TealOp(None, Op.bnz, indexToLabel(trueIndex)))
-                continue
-
-            if trueIndex == i + 1:
-                references[falseIndex] += 1
-                code.append(TealOp(None, Op.bz, indexToLabel(falseIndex)))
-                continue
-
-            references[trueIndex] += 1
-            code.append(TealOp(None, Op.bnz, indexToLabel(trueIndex)))
-
-            references[falseIndex] += 1
-            code.append(TealOp(None, Op.b, indexToLabel(falseIndex)))
-        else:
-            raise TealInternalError("Unrecognized block type: {}".format(type(block)))
-
-    teal: List[TealComponent] = []
-    for i, code in enumerate(codeblocks):
-        if references[i] != 0:
-            teal.append(TealLabel(None, indexToLabel(i)))
-        teal += code
-
-    return teal
 
 def verifyOpsForVersion(teal: List[TealComponent], version: int):
     """Verify that all TEAL operations are allowed in the specified version.
@@ -137,7 +51,7 @@ def verifyOpsForMode(teal: List[TealComponent], mode: Mode):
             if not op.mode & mode:
                 raise TealInputError("Op not supported in {} mode: {}".format(mode.name, op))
 
-def compileTeal(ast: Expr, mode: Mode, *, version: int = DEFAULT_TEAL_VERSION) -> str:
+def compileTeal(ast: Expr, mode: Mode, *, version: int = DEFAULT_TEAL_VERSION, assembleConstants: bool = False) -> str:
     """Compile a PyTeal expression into TEAL assembly.
 
     Args:
@@ -146,12 +60,18 @@ def compileTeal(ast: Expr, mode: Mode, *, version: int = DEFAULT_TEAL_VERSION) -
         version (optional): The TEAL version used to assemble the program. This will determine which
             expressions and fields are able to be used in the program and how expressions compile to
             TEAL opcodes. Defaults to 2 if not included.
+        assembleConstants (optional): When true, the compiler will produce a program with fully
+            assembled constants, rather than using the pseudo-ops `int`, `byte`, and `addr`. These
+            constants will be assembled in the most space-efficient way, so enabling this may reduce
+            the compiled program's size. Enabling this option requires a minimum TEAL version of 3.
+            Defaults to false.
 
     Returns:
         A TEAL assembly program compiled from the input expression.
 
     Raises:
         TealInputError: if an operation in ast is not supported by the supplied mode and version.
+        TealInternalError: if an internal error is encounter during compilation.
     """
     if not (MIN_TEAL_VERSION <= version <= MAX_TEAL_VERSION) or type(version) != int:
         raise TealInputError("Unsupported TEAL version: {}. Excepted an integer in the range [{}, {}]".format(version, MIN_TEAL_VERSION, MAX_TEAL_VERSION))
@@ -188,6 +108,11 @@ def compileTeal(ast: Expr, mode: Mode, *, version: int = DEFAULT_TEAL_VERSION) -
     for index, slot in enumerate(sorted(slots, key=lambda slot: slot.id)):
         for stmt in teal:
             stmt.assignSlot(slot, index)
+    
+    if assembleConstants:
+        if version < 3:
+            raise TealInternalError("The minimum TEAL version required to enable assembleConstants is 3. The current version is {}".format(version))
+        teal = createConstantBlocks(teal)
 
     lines = ["#pragma version {}".format(version)]
     lines += [i.assemble() for i in teal]